refactor: change port's type to int
The "Endpoint" struct is expecting an int, and not an "int64" as I
defined in the "EndpointCreateRequest" struct.

+++ service/endpoint_validation.go
@@ -26,7 +26,7 @@ var validAvailabilityStatuses = []string{"", model.Available, model.Unavailable}
 
 const (
 	defaultScheme    = "https"
-	defaultPort      = int64(443)
+	defaultPort      = 443
 	defaultVerifySsl = true
 	maxFqdnLength    = 255 // RFC2181
 	maxLabelLength   = 63  // RFC2181
